signer: generalize the txn hashing in the signer
This is needed since the LegacyTxn after EIP-155, needs the chain_id
which isn't part of the transaction data. Also, technically speaking the
right signature depends on the fork and only the signer knows that.

So, moving the txn hashing for the signer is the right thing now and
probably for the future too.